(#15852) Allow rc in Puppet and Facter version strings
Previously putting rc in the Puppet version string would fail, mainly because
of a test that verified that the version only contained 3 dot separated
integers. This commit updates the Facter dependency to correctly use 2.0.0-rc
as the minimum version, as 2.0.0-rc is less than the previously required 2.0.0
as specified by SemVer. It also removes the test that verified the Puppet
version was a 'valid' 3 dotted integer version. That test is redundant, as the
test that follows verifies that the version string is valid SemVer.