(#11870) Pass a list of Arrays to multiple_yields
Previously, I was passing a list of packages to mocha's multiple_yields,
which on ruby 1.9.3-p0 results in an unexpected call to the `to_a`
method on the mock object, but on 1.8.7 and later versions of 1.9.3 does
not.

The real issue is that I'm not calling multiple_yields correctly:

  Note that each +parameter_group+ should be an Array representing the
  parameters to be passed to the block for a single yield.

This commit changes it to pass a list of Arrays, where each element in
the array is yielded in turn (as opposed to the Array itself).

This commit also adds names to the mock objects to assist with debugging
when the expectation fails.