(#11870) Ignore uninstall keys that don't exist
Previously, we assumed the 4 combinations of (HKLM or HKCU) and (32 or
64) registry paths always existed. When running puppet in an interactive
logon session, this assumption appears to hold. However, when puppet
runs as a service (under LocalSystem), the assumption fails as the
registry key `Software\Microsoft\Windows\CurrentVersion\Uninstall`
typically does not exist within the HKEY_CURRENT_USER hive. In that
case, we would fail to enumerate the set of installed products.

This commit changes the enumeration code to ignore errors that occur
when the Uninstall key for a particular combination does not exist. This
corresponds the Windows error code ERROR_FILE_NOT_FOUND[1]

However, if an exception occurs for other reasons, e.g.
ERROR_ACCESS_DENIED, then we re-raise the exception and fail.

Also, the inner `open` call should never fail with ERROR_FILE_NOT_FOUND
because we get a list of subkeys for the Uninstall key, and then open each
subkey in turn.

[1] http://msdn.microsoft.com/en-us/library/windows/desktop/ms681382(v=vs.85).aspx