(Maint) Fix rspec options file
The options file for spec testing is wrong without this patch applied.
spec.opts is no longer read by RSpec, instead .rspec in the root of the
repository is read.

This patch removes the old and unused file and replaces it with .rspec
which enables progress format, color, and randomized tests.