(#15464) Make Puppet usable as a Gem from source
Without this patch it is difficult to use Puppet as a rubygem from a
source checkout.  This is a problem because tools like Bundler will
install the gem from rubygems.org if it believes the gem is not
installed.

This patch fixes the problem by adding a `puppet.gemspec` file and
information in README_DEVELOPER about how to make the repository clone
available as a gem using the GEM_PATH and two symbolic links.