(#8200) munge the ensurable value.
We modify munge so that the error "changing from 'running' to 'running'"
is avoided on `zone apply ...` on a zone that is actually in sync.
This results from comparison of string value that we get from manifest
(because we override newvalue) to the symbol generated by the resource.
This is a temporary measure, and should be tackeled in #15886