(#8200) Munge the ensurable value
Previously, the ensure property was always out-of-sync, because we were
comparing the current symbol value against the should string value. This
occurred, because the type's ensurable methods redefine
`Puppet::Parameter.newvalue`, which has the side-effect of no longer
converting string should values to symbols.

This commit converts the should value to a symbol by defining a munge
method. It is a temporary measure until the real issue can be fixed in
15886.