(#8200) Change path from parameter to property
Previously path was a parameter which resulted in it not being included in
the resources listing. However since this is a required attribute, we
change it to a property instead.

We also remove the path validation from global validate and instead add
this check in each place that path is required.