(#8200) change path from param to property.
Previously path was a param which resulted in it not being included in
the resources listing. However since this is a required parameter, we
change it to a property instead.

We also remove the path validation from global validate and instead add
this check in each place that path is required.