(#15886) commit the complete commands at the end.
Previously, the provider was applying the setconfig commands as soon as
it was called. This caused problems when the commands that were being called
needed to be called in a transactional manner
(i.e eth,exclusive -> eth:ip,shared). This uses flush to pass the required
commands at once through a pipe to zonecfg rather than do it one at a time.