(#15464) Exclude gem command from bundler
This patch simply excludes the `gem` command from the dependency
handling of Bundler with the `rubygems-bundler` gem installed.

This patch also updates the README to describe the resolution to the
spec test failures we've been seeing when Puppet contains a Gemfile in
the project root.

To summarize: Use of RVM is considered dangerous

We're able to work around the problem, but the hijacking of shell
builtin commands is a maintenance burden we're not willing to shoulder,
hence RVM being considered a dangerous tool.