(#15975) Proxy the request_get block to Net::HTTP
The fix in a40639a is insufficient because it silently drops the block
passed to the request_get block.  This is a problem because it causes
file content to be silently dropped and truncated to zero bytes.

This patch fixes the problem by proxying the get request block along to
the underlying implementation in Net::HTTP.