(#15975) Add missing methods to new HTTP::Connection class
Some of the callers who were getting an instance of the
new HTTP::Connection class via HttpPool#http_instance
were expecting it to provide the newer request_* methods
that are available on a Net::HTTP instance, so that (e.g.)
they can pass in a block to be executed against the response.

Prior to this commit the HTTP::Connection object did not
expose those methods; this adds them, which should
fix some failing acceptance tests.