(Maint) Fix Puppet::Util.execute warning
Without this patch applied the end user receives a nasty deprecation
warning when running `puppet agent` with a `puppet.conf` that configured
the node terminus to exec as in the following example:

    [main]
    node_terminus = exec
    external_nodes = $confdir/enc.rb

This a problem because we shouldn't issue warnings to the users about
our own use of our own code.

This patch fixes the problem by proxying the execute method from the
object that is self when the find method is called from the child class
using super.