(#15797) Disable RedHat services in all run levels
Without this patch applied the `chkconfig <name> off` command is used to
disable services.  This command operates on run levels 2,3,4 and 5 by
default.  This behavior is a problem because the `chkconfig <name>
reset` method may turn the service on in run levels 0,1 or 6.

This patch fixes the problem by explicitly disabling the service in all
run levels.