(#15886) Refactor state machine
The zone type has an internal statemachine that ensures that the state
transition is linear from configured -> installed -> running.
Previously, the state machine was intertwined with the code of zone
type. This commit extracts the statemachine so that thestatemachine is
in its own class

In order to remove future confusion, we also rename newvalue to seqvalue
and ensure that the newvalue is called from within the seqvalue. The
seqvalue is so named to indicate that the order of declaration of the
seqvalues determine the order of progression of states in the state machine.
Since we call parent newvalue from inside the seqvalue, we can now avoid using
munge which was added for this reason.

- Added test cases for fsm