(#15886) remove ZoneConfig and ZoneMultiConfig
Remove the zone specific properties from types, and move their handling to the provider.
This is done so that we can avoid stepping on the toes of puppet core and override methods
that have default behaviors except when necessary. The zone type is also more cleaned up
so that it can serve as a base for other container implementations.

Type: the properties that earlier relied on ZoneMultiConfigProperty are now moved to
Puppet::Property::List. The methods should and insync? are added to override the behaviors in
the parent methods. The should method in parent joins the array values together before
comparison. However, we do not want this behavior, and instead want it to remain as an array.
Hence we override should. We also override insync?  which uses should in parent for the same
reason.

Provider: the def_prop and def_multiprop now handles the definition of properties. They
add the methods prop(), prop=(), and prop_conf() to each properties that the provider provides.

The prop_conf() method is needed for getting the text to send to zonecfg on configure.