(#15193) Document how to use the static compiler
Without this patch the magic nature of the `Filebucket[puppet]` resource
and the expected behavior of the puppet agent obtaining file content
from the server is not documented anywhere.  This is a problem because
it is not clear how to configure the agent in a manner that does not
produce errors when using the static compiler.

This patch fixes the problem by providing a more in-depth description of
the high-level approach the static compiler takes.  In addition, the
specific configuration of `fileserver { puppet: server => $server, path
=> false }` is documented in the README_DEVELOPER.