(Maint) Tighten up assertion to find the cause of a flicker
The test for the message emitted for calling an incorrect action on a
face sometimes fails because it sees "Could not run Puppet configuration
client: Not a directory - /dev/null/state/agent.pid" instead. This
changes the way the assertion is done so that the test failure will
provide more information about what else is in the log at the time of
the failure, which might help in tracking down what is going wrong.