(#16112) Add Augeas clearm command support
Without this patch applied, When trying to use clearm in augeas provider
I get the following error:

    err: /Augeas[boot_grub_menulst_kernel_param_timekeeping_use_tsc]:
    Could not evaluate: Unknown command clearm

This patch fixes the problem by using the Ruby augeas bindings to test
if the clearm command is supported.  If it is not, the catalog run
fails, otherwise it continues.

This support does not take the form of a new parameter.  Instead, the
patch follows the existing augeas convention of using the `changes`
parameter.