(Maint) Remove dependency on Facter 2.0
Previously a hard dependency had been made on Facter 2.0. As the release
has neared it turns out that Facter 2.0 is not ready for a release and
Puppet really does not need any functionality that is present in that
version of Facter. This removes this dependency, which allows Puppet to
use the existing Facter 1.6.x series and thereby decouples the release
of the two systems.