(#15193) Add spec test coverage for static compiler
Without this patch there is no test coverage for the static compiler
terminus.  This is a problem because internal changes to the indirector
break the expected behavior of the static compiler and we have no
automated way to discover the broken behavior.

This patch adds spec test coverage for the static compiler's find
method.  This covers the most common use case of the static compiler
where by file resources with source parameters are modified in the
resource catalog.

The patch does not include test coverage for recursive file resources.

The approach taken stubs the underlying compiler terminus and produces a
fake resource catalog for the static compiler to operate on.  The
fileserver metadata API is also mocked to break the dependency on a real
fileserver.