Document the different forms of catalogs to help test writers
Without this patch it was not clear to me that the static compiler
operates primarily on a resource catalog.

This was a problem because I spent considerable amounts of time trying
to figure out the difference between `resource[:source]` returning
`"puppet:///..."` inside of
`Puppet::Resource::Catalog::StaticCompiler#find` when running `puppet
apply` and `resource[:source]` returning `["puppet:///..."]` (an array)
when inside of `#find` from the spec test.

When testing catalog terminii the resource form of the catalog should be
created instead of the RAL form.

This patch addresses the problem by providing a description of the
difference between a resource catalog and a RAL catalog and when it is
appropriate to use each.