Move all packaging artifacts from conf to ext
This commit moves all packaging artifacts out of the conf dir and into
the ext dir. This is generally a move to align packaging across all
puppetlabs products and make the directories more sensible, e.g. conf
should have config files. The redhat spec and install.rb are updated
for the path change.

Signed-off-by: Moses Mendoza <moses@puppetlabs.com>