(#15591) Skip test that causes segfault on Windows
Previously, this test would segfault on Windows with ruby 1.8.7 p370, but
not p334. The test can never work on Windows as currently written
(Kernel.fork). And we can't mark it as pending, because rspec attempts to
run the test, expecting either a failure or exception, but ruby segfaults.
So we have to skip entirely on Windows.

Note in the merge up to 3.x, this spec file has been removed.