Maint: Use pending block
This commit removes the `fails_on_windows` filter in favor of a pending
block. This way we document what tests don't currently pass on Windows,
and it ensures that the test fails or raises. If the test runs without a
problem, the test will be reported as a failure, so that we know the test
is no longer pending.