(Maint) Stop watchr from reading non-existent spec.opts
Without this patch the watchr auto-testing script attempts to read the
spec.opts file.  This is a problem because the file no longer exists in
the repository which causes watchr to fatally error out.

This patch fixes the problem by simply removing the code replacing it
with a comment to use a personal ~/.rspec file instead of the project
rspec options file.