(#15193) Fix windows failing on unix style paths
Without this patch the windows spec tests are failing because the
resource contains File resources with *nix style absolute paths.  These
paths are not considered absolute on windows, so a simple
File.expand_path converts the relative path to an absolute path on
windows.

This patch should restore the build to green.