(#15731) Add clarifying comments to host.rb
Prior to this commit it was unclear that the existing :fingerprint
element of the pson hash was still around for backwards compatibility
for as of yet unknown external consumers of the existing api. This
commit makes it clear that that's why it's still around and that we
don't have any known internal consumers of that api.