(#8714) Don't follow symlinks in SELinux FS detection
When detecting the filesystem that a target file lives on, we used to follow
symbolic links.  This is actually a problem - the SELinux change applies to
the link, not the target of the link.

This rewrites the code to handle this more cleanly, following the logical tree
up to locate the record in the set of mountpoints.

Signed-off-by: Daniel Pittman <daniel@puppetlabs.com>