(#7962) Added rate-limiting logger utility
`Puppet::Util::Log::RateLimitedLogger` is a class that includes the
standard logging module and so responds to `info`, `warning`, etc. It
uses log messages themselves as keys to a hash that records the last time
they were logged and at which level. When a message is logged, it first
checks the record and ignores the call if the given time interval has
not yet elapsed since the last time it was logged.