AST Scope variable names must be strings.
The Parser AST scope had an embedded assumption that variable names were
always strings, but didn't actually enforce that.  A consequence of this is
that a couple of tests violated that agreement.

This adds an internal error report when someone tries the wrong thing, and
cleans up the handful of tests that failed.

Signed-off-by: Daniel Pittman <daniel@puppetlabs.com>