(#15756) Do not allow functions called without Array
Proper calling of functions from templates, and anywhere else in ruby
code that calls puppet functions, requires that the function's arguments
are passed as an Array. Previously this was never enforce and some
functions would end up working when not called with an Array and other
functions would fail.

This adds a type check to ensure that functions must always be called
consistently.