(#16027) Restore function context to backtraces
The changed implementation of `pretty_backtrace` was indeed prettier, but it
also lost information: the function context was dropped from the trace.

Since this is a debugging tool, and knowing the method tree is quite valuable
to understanding a problem, this restores it - and any extra context that Ruby
later decided it should add to the picture.

Signed-off-by: Daniel Pittman <daniel@puppetlabs.com>