Eliminate asynchronous StoreConfigs entirely.
PuppetDB provides a better, faster, and more effective solution to the problem
that asynchronous StoreConfigs was tackling - the inefficiency of ActiveRecord
no longer matters.

Eliminating this allows us to remove a bunch of queue related code and clean
up various hacks in the system.  Anyone who did use this should either move
to synchronous StoreConfigs, or better, deploy PuppetDB.

Signed-off-by: Daniel Pittman <daniel@puppetlabs.com>