(#14766) Ensure that splay applies to all agent runs.
When the event loop was rewritten, one change was to base the next event off
the start time of the previous event, not the completion time.  This is great
in terms of consistency - we no longer drift around the clock based on random
completion times.

The unfortunate side-effect was that splay stopped working: because it would
sleep at the start of the run, it would cause extra delay in the first agent
invocation but have no effect on subsequent runs - they ran strictly every
`runinterval` from the time the process started.

To fix this the knowledge about splay time is moved out of the agent itself
and into the scheduler, which can delay the first and the subsequent runs.

It also separates out the decision making about the next run schedule from the
actual scheduler loop, making the whole thing easier to understand in
isolation, as well as to test.

Signed-off-by: Daniel Pittman <daniel@puppetlabs.com>