(#15756) Test behavior of function loading not mocks
The tests used to test that functions were created by
asserting the mechanism by which this happened. This made them brittle
in the face of changing implementation. This changes the tests to make
assertions about the produced module and should be much more robust
against changing implementation and more expressive of the behavior of
defining functions.