(acceptance) utils: cleanup - refactor zone utils so that we check before cleanup
Previously, we were duing zoneadm uninstall,delete, destroy etc, without
regard to the actual state. However, zfs destroy -r has a nasty tendency
to hang some times when given an non-existant fs. Moreover, the deletion
of phantom resources is confusing when inspecting the output. This patch
refactors zoneutils so that we do not try to remove non-existant resources.