(acceptance) disable zone tests that require installation
Zone tests that require installation often results in the system under tests to
hang until the installation is completed. The problem is worse in solaris11 which
has added a lot more packages to the default small-server-install. These tests
are disabled so that we can avoid ~2 hour delay on running per install command.