(acceptance) fix zones: ensure that we pass a vallid interface name, fix name typo, sizes
Previously, the interface name that was being passed for creating zones
was specific to the virtual machine that the tests were developed on. This
patch changes that behavior to use a more generic name. (Note that the
valid interface name is documented nowhere. The only reference I could find
was the c file that implemented the function.)