(#13435) Use digest algorithm from CSR for fingerprint
Since the digest algoritm changed between Puppet 2.7 and 3.0 from MD5 to
SHA256, there was a problem when submitting a CSR from a 2.7 agent to a 3.0
master. The agent would display the fingerprint using MD5, and the master
would display the SHA256 fingerprint, making it impossible to verify using the
normal puppet commands. This makes the default digest algorithm for each CSR
the algorithm that was used to generate the signature. This will be MD5 for a
CSR generated on 2.7, and SHA256 for a CSR generated on 3.0. When showing a
fingerprint puppet will now always print the digest algorithm that was used.
All fingerprinting commands should also accept a --digest option to specify
the digest.

It also significantly refactors and unifies the code used to generate and
display fingerprints. It introduces a new class, Puppet::SSL::Digest, to
encapsulate digest data, and uses this everywhere a digest is needed.