Remove asc file a source for RPM builds
We used to require the asc file for RPM builds of puppet.  This was not
consistant with our packaging of any other platform and raised the barrier to
entry if you were trying to build your own puppet package.

This commit removes the requirements of signing the tarball for building an RPM
package.  Tarball will still be signed prior to official release from Rel-Eng
team at Puppet Labs.