rpm specfile improvements for ruby19
This commit fixes up puppet packaging to make Puppet 3.0 (and higher, AKA
Telly) require ruby19 on EL5 systems. This is in an effort to deprecate ruby
1.8.5 (the stock ruby ) on EL5.  Puppet Labs has packages for Ruby19 that will
be submitted to EPEL and also available on yum.puppetlabs.com for consumption
in the EL5 branch.

Note: this means that now puppet 3.x will *not* build using stock system ruby
on EL5.

It also fixes up the spec file to be more
compatible with Fedora packaging guidelines.

Signed-off-by: Michael Stahnke <stahnma@puppetlabs.com>