(#16425) ensure that first time install of speciifc version works
Previously if a specific version was requested, and it was the first time, the package installation
errored out due to passing the current version :absent to version compare. This patch verifies that
we do not pass absent to version compare if we know it is absent in the system currently.